Add shared PuzzleState model and integrate
