+++ PuzzleAM/Hubs/PuzzleHub.cs
@@ -1,23 +1,9 @@
 using System.Collections.Concurrent;
 using System.Linq;
 using Microsoft.AspNetCore.SignalR;
+using PuzzleAM.Model;
 
 namespace PuzzleAM.Hubs;
-
-/// <summary>
-/// Represents the position of a puzzle piece on the board.
-/// </summary>
-public record PiecePosition(int Id, float Left, float Top, int? GroupId);
-
-/// <summary>
-/// Maintains the current state of a puzzle session.
-/// </summary>
-public class PuzzleState
-{
-    public string ImageDataUrl { get; set; } = string.Empty;
-    public int PieceCount { get; set; }
-    public ConcurrentDictionary<int, PiecePosition> Pieces { get; } = new();
-}
 
 public class PuzzleHub : Hub
 {
